Elliott can use a default advisory if defined in ocp-build-data
To make bug sweeping happen more frequently and automatically we are
adding this little bit of functionality. If a value is defined in
ocp-build-data for .advisories.<TYPE> (type=rpm, image, security) then
you can have elliott attach bugs to that advisory by selecting it on
the command line.

We have played around with implementing logic to automatically
determine the right advisory for this but it hasn't come to fruition
yet. Instead we're going to try this little workaround instead.