ART-2170: do not count shipped images as missing
We no longer (necessarily) rebuild and ship all images each week, which means part of the payload may not be new and won't be present in the advisory we ship. In this situation elliott verify-payload should not count shipped images as missing.

This PR also show images already attached to advisories (useful when preparing a release but the previous release is on the fly) and introduces asyncio to boost the performance.