[ART-1054] introduce verify-attached-operators (#161)
Verify attached operator manifest references are shipping or already shipped.

Takes a list of advisories that may contain operator metadata/bundle builds or image builds that are shipping alongside. Then determines whether the operator manifests refer only to images that have shipped in the past or are shipping in these advisories. An error is raised if there are no manifest builds attached, or if any references are missing.