verify-attached-bugs: Don't fail on RHBA
When running `verify-attached-bugs` for RHBA which were previously an
RHSA, a lot of CVE flaw related checks will fail because Errata has some
records (like CVE Names and CVE package exclusions) remained.
We need to skip those checks if the advisory is RHBA.

This PR also adds a check to raise an error if the advisory is not RHSA and
first-fix flaw bugs are attached.