Halt if rpm CVEs don't have corresponding build
Throw an error when there are CVE tracker bugs that cannot get attached
because the component that they are about is not getting shipped.

Currently an event like this will go unnoticed, as the log is not
scrutinized.

The can be bypassed by either fixing the issue, or by exclkuding the
relevant bugs from the assembly:

`releases.<name>.issues.exclude: [1234]`