Fix version detection error on pip installations
Fix the following error when invoking `elliott --version` on an
installation without the source git repository:

```
LookupError: setuptools-scm was unable to detect version for /usr/local/lib/python3.11/site-packages.
```

Steps to reproduce:

```sh
pip install .

elliott --version
```

This can be solved by introducting setuptools_scm to write a file
containing the to-be-installed version.