verify-attached-bugs: Fix Target Version not set error
A validation in `verify-attached-bugs` is checking if all blocking bugs
in the previous minor version are already addressed (in VERIFIED or
later state). However, I noticed the following error when running it for
4.13 rc.3:

```
  File "/mnt/workspace/jenkins/working/cd-builds_build_promote-assembly@2/art-tools/elliott/elliottlib/cli/verify_attached_bugs_cli.py", line 176, in validate
    blocking_bugs_for = self._get_blocking_bugs_for(non_flaw_bugs)
  File "/mnt/workspace/jenkins/working/cd-builds_build_promote-assembly@2/art-tools/elliott/elliottlib/cli/verify_attached_bugs_cli.py", line 367, in _get_blocking_bugs_for
    next_target = any(is_next_target(target) for target in bug.target_release)
  File "/mnt/workspace/jenkins/working/cd-builds_build_promote-assembly@2/art-tools/elliott/elliottlib/bzutil.py", line 231, in target_release
    raise ValueError(f'bug {self.id} does not have `Target Version` field set')
ValueError: bug OCPBUGS-6518 does not have `Target Version` field set
```

This error occurs if one of the `depends_on` bugs has no `Target Version` field set.
Calling `jira_bug.target_release` will trigger a `ValueError` if the
`Target Version` field of `jira_bug` object is not set.

In case a "depends_on" bug has no `Target version`, we shouldn't
consider it to be a blocking bug. So this PR will ignore that kind of
exception.